+++ packages/terra-navigation-side-menu/CHANGELOG.md
@@ -6,6 +6,9 @@ Unreleased
 ### Changed
 * Remove use of componentWillReceiveProps
 
+### Added
+* Add optional toolbar prop that displays below the side menu action header
+
 1.18.0 - (December 5, 2018)
 ------------------
 ### Changed