+++ packages/terra-navigation-side-menu/CHANGELOG.md
@@ -4,6 +4,8 @@ Changelog
 Unreleased
 ----------
 
+1.19.0 - (January 2, 2019)
+------------------
 ### Fixed
 * Updated Data-menu-item attribute value of Menuitem to Unique key value